@@ -1,11 +1,11 @@
 name: rapidoreach
 description: Monetize your users through rewarded surveys!
-version: 1.0.9
+version: 2.0.0
 homepage: https://www.rapidoreach.com
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
-  flutter: ">=2.5.0"
+  sdk: ^3.8.0
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
@@ -14,7 +14,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
+  flutter_lints: ^6.0.0
 
 flutter:
   plugin:
@@ -24,5 +24,3 @@ flutter:
         pluginClass: RapidoReachPlugin
       ios:
         pluginClass: RapidoReachPlugin
-
-
